consolidate backend interface in one file
